Move handleRpcError to its own file
